Day 13
Part 2 reduces to computing the intersections of sets of the form aZ+b,
where a and b are integers and Z is the set of all integers. I've done
this before in my previous life (see
https://github.com/DMRobertson/thompsons_v/blob/master/thompson/number_theory.py)
and thought that porting it over to rust would be a doddle.

In actual fact, I was tired and made lots of small mistakes, plus I'd
forgotten all the number theory going on behind the scenes and had to
rederive it. I'm not sure how comprehensible the result is. But it does
seem to work!