Day 17
Initially tried to do this with multidimensional arrays and lots of
index manipulation. I found it quite fiddley to deal with referring
from one iteration of the grid to the previous one. Lots of subtractions
by one that were causing usize underflows.

So ditch that. Try and keep it simple and use a hashmap for storage. It
worked and was a lot easier to implement. But Part Two was a bit
sluggish. Poor access locality? Or just lots of computation?

It works, but I'm not very proud of this one.