Day 18
Happier with this one. I don't fully grok the parsing stuff---had to
consult some googling here. But I suppose the point is if you were going
to do this, you'd probably define a grammar and use that to generate a
tokenizer/lexer and a parser, particularly to handle all the precendence
stuff.

Evaluation was a doddle by comparison!

I didn't fully understand everything I was going with the Box<>es.
It all seemed to "just work" and move the pointers around like I'd hoped
it would. But it feels a little bit like magic. I don't think I could
describe the implementation in the same way that I could describe unique_ptr.