Day 20: part one
Didn't work with the proper input data. So try it on the reversed list
of tiles on the test data. It fails. Urgh. A pain to track down what was
wrong. We weren't removing ids from the Compositions set of ids if we
had to move from one child to another of the DFS.

Not sure if that was always the problem or if it was the problem after I
made a bunch of changes today. But we got there.

Part 2 tomorrow?